move COLLECTION_NAME to vectordb dir
