+++ context_chat_backend/controller.py
@@ -45,8 +45,8 @@ def _(query: str | None = None):
 @app.get('/vectors')
 @enabled_guard(app)
 def _(userId: str):
-    from chromadb import ClientAPI
-    from .utils import COLLECTION_NAME
+	from chromadb import ClientAPI
+	from .vectordb import COLLECTION_NAME
 
     db: BaseVectorDB = app.extra.get('VECTOR_DB')
     client: ClientAPI = db.client
@@ -62,7 +62,7 @@ def _(userId: str):
 @enabled_guard(app)
 def _(userId: str, keyword: str):
 	from chromadb import ClientAPI
-	from .utils import COLLECTION_NAME
+	from .vectordb import COLLECTION_NAME
 
 	db: BaseVectorDB = app.extra.get('VECTOR_DB')
 	client: ClientAPI = db.client
